Clean Scripts

+++ Assets/Script/Controller_2Stage.cs
@@ -37,16 +37,16 @@ public class Controller_2Stage : MonoBehaviour {
         {
             interaction.value2 = !interaction.value2;
         }
-        /*
-        if (SerialHandler.getSensorDown(interaction.sensorTrigger1)
+        
+        if (SerialHandler.getSensorDown((int)interaction.sensorTrigger1))
         {
             interaction.value = !interaction.value;
         }
-        if (SerialHandler.getSensorDown(interaction.sensorTrigger2)
+        if (SerialHandler.getSensorDown((int)interaction.sensorTrigger2))
         {
             interaction.value2 = !interaction.value2;
         }
-        */
+        
 
         if (state == 0)
         {
